% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertselect_removal.R
\encoding{UTF-8}
\name{convertselect_removal}
\alias{convertselect_removal}
\title{
Creating input date files for Bayesian population dynamics models 
}
\description{
\code{convertselect_removal} is used by \code{\link{popbay_model}} to select and convert the variables required for hierarchical Bayesian population dynamic models that account for individual detection errors.
}
\usage{
convertselect_removal(data, time, start, time.interval,
  type, period, para.bas)
}
\arguments{
  \item{data}{
a data.frame containing the variables used in the model and possible additional items (variables order does not matter).
}
  \item{time}{
a numeric vector of dates included in modelling.
}
  \item{start}{
a numeric that defines the starting date for estimating average rates.
}
  \item{time.interval}{
a numeric that defines the time interval between two consecutive dates.
}
  \item{type}{
a character that defines the modelling type (values are A, B, AB, O, OA, OB or OAB).
}
  \item{period}{
a numeric that defines the time interval for estimating intermediate average rates, if NA the required variables to estimate the intermediate average rates are not included in the model variables list.  
}
  \item{para.bas}{
a logical, if FALSE the required variables to estimate the regional average rates are not included in the model variables list.
}
}
\value{
\code{convertselect_removal} returns a list containing the list of model variables (\code{data.jags}), the vector of site identifiers (\code{sta}), the vector of region identifiers (\code{bas}) and the vector of protocol identifiers (\code{pro}).
}
\seealso{
\code{\link{popbay_model}}, \code{\link{writemodel_removal}}
}
\examples{
\donttest{data(riverfish)
varlist <- convertselect_removal(riverfish,
  time=sort(unique(riverfish$date)),
  start=min(riverfish$date),
  time.interval=1,
  type="A",
  period=NA,
  para.bas=TRUE)
}
}
