% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modparameters.R
\name{get_modparameters}
\alias{get_modparameters}
\title{Description of modelling function parameters}
\usage{
get_modparameters(fun, parameter)
}
\arguments{
\item{fun}{[optional, character] name of the modelling function}

\item{parameter}{[optional, character] name(s) of the parameter(s)}
}
\description{
Gives description of the parameters of modelling functions.
}
\examples{
# Get description about all parameters of mod_popoccup function
get_modparameters(fun = "mod_popoccup")

# Get description about some parameters
get_modparameters(parameter = "z")
get_modparameters(parameter = c("z","N","B"))
}
