% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverfish.R
\docType{data}
\name{indriverfish}
\alias{indriverfish}
\title{Fish individual measures}
\description{
This dataset gives the individual lengths and weights of four freshwater fish species. 
}
\usage{indriverfish}
\format{
  \code{indriverfish} is a data frame with 3 variables:
  \describe{
    \item{\code{taxa}}{species common names}
    \item{\code{length}}{individual length in mm (total length)}
    \item{\code{weight}}{individual weight in g}
   }
}
\source{
\code{indriverfish} is based on data extracted from the national ASPE database managed by the
French Biodiversity Agency (OFB).
}
\keyword{datasets}
\keyword{ASPE}