% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_datamodel.R
\name{int_datamodel}
\alias{int_datamodel}
\title{Title}
\usage{
int_datamodel(
  df,
  occup,
  grow,
  modenv,
  modenvG,
  timestep,
  period,
  var_id,
  var_tmp,
  var_tax,
  var_pres,
  var_reg,
  var_guild,
  var_cnt,
  var_wei,
  var_surf,
  var_envO,
  var_envP,
  var_envC,
  var_grow
)
}
\arguments{
\item{var_grow}{}
}
\description{
Title
}
