% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_transformsummary.R
\name{int_transformsummary}
\alias{int_transformsummary}
\title{Title}
\usage{
int_transformsummary(mcmc_summary, datamodel, var_id, var_tmp, var_tax, period)
}
\arguments{
\item{period}{}
}
\description{
Title
}
