% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_convertquali.R
\name{mef_convertquali}
\alias{mef_convertquali}
\title{Converts quantitative variables into qualitative variables}
\usage{
mef_convertquali(df, var_quali)
}
\arguments{
\item{df}{data frame}

\item{var_quali}{argument}
}
\value{
df data frame
}
\description{
Converts quantitative variables into qualitative variables
}
\examples{
data(riverfish)
mef_convertquali(riverfish, var_quali="flow")
}
