% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lenwei.R
\name{mod_lenwei}
\alias{mod_lenwei}
\title{Bayesian modelling for the estimation of length-weight relationships}
\usage{
mod_lenwei(
  df,
  var_len,
  var_wei,
  var_tax = NULL,
  n_chain = 3,
  n_iter = 102000,
  n_thin = 100,
  n_burnin = 2000
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{var_len}{[character] variable name with individual lengths of fish (must be numeric without missing values)}

\item{var_wei}{[character] variable name with individual weights of fish (must be numeric without missing values)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{n_chain}{[optional, numeric] number of MCMC chains}

\item{n_iter}{[optional, numeric] number of iterations per MCMC chain}

\item{n_thin}{[optional, numeric] backup interval for MCMC samples}

\item{n_burnin}{[optional, numeric] number of initial iterations to discard per Markov chain}
}
\value{
\code{mod_lenwei} returns a list with the summaries of main statistics associated with the estimation of the length-weight relationship parameters (\code{mcmc_summary}) and the MCMC chains (\code{mcmc_chain}).
}
\description{
Fits Bayesian models to individual measures for estimating length-weight relationships of one or more taxa (\code{Weight = a . Length^b}).
}
\examples{
\dontrun{
## Length-weight relationships for all taxa             /
mcmc.out <- mod_lenwei(indriverfish,
                       var_len = "length",
                       var_wei = "weight",
                       var_tax = "taxa")

## Length-weight relationship fo eel 
df <- indriverfish[indriverfish$taxa \%in\% "eel",]
mcmc.out <- mod_lenwei(df,
                       var_len = "length",
                       var_wei = "weight")
}
}
