% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popgrowAlt.R
\name{mod_popgrowAlt}
\alias{mod_popgrowAlt}
\title{Alternative population growth modelling in a Bayesian hierarchical framewor}
\usage{
mod_popgrowAlt(
  df,
  var_id,
  var_tmp,
  var_tax = NULL,
  var_cnt = NULL,
  var_wei = NULL,
  var_surf = NULL,
  var_pro = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 102000,
  n_thin = 1000,
  n_burnin = 2000
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{var_id}{[character] variable name with sampling location identifiers (must not contain missing values)}

\item{var_tmp}{[character] variable name with continuous sampling times (must be numeric without missing values)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{var_cnt}{[optional, character] variable name with taxa abundances (must be numeric)}

\item{var_wei}{[optional, character] variable name with taxa biomasses (must be numeric)}

\item{var_surf}{[optional, character] variable name with quantitative sampling effort (e.g. sampling surfaces). If supplied, the model takes into account variabilities in sampling effort from one time to the next (must be numeric without missing values)}

\item{var_pro}{[optional, character] variable name with sampling protocol (must not contain missing values)}

\item{var_reg}{[optional, character] variable name(s) with spatial level identifiers. If supplied, estimates of occupancy changes rates at intermediate spatial levels are added to the model. Otherwise, occupancy change rates are only estimated across all locations}

\item{var_guild}{[optional, character] variable name(s) with taxa guilds identifiers. If supplied, estimates of guild occupancy changes rates are added to the model}

\item{period}{[optional, numeric] list of vectors with start and end times for estimating intermediate average growth rates}

\item{timestep}{[optional, numeric] interval between two consecutive times to account for regular gaps in time series in the rate estimates}

\item{save_parameters}{[optional, character] name(s) of the parameter(s) to save (by default, all model parametes are saved)}

\item{n_chain}{[optional, numeric] number of MCMC chains}

\item{n_iter}{[optional, numeric] number of iterations per MCMC chain}

\item{n_thin}{[optional, numeric] backup interval for MCMC samples}

\item{n_burnin}{[optional, numeric] number of initial iterations to discard per Markov chain}
}
\details{
\code{mod_popgrowAlt} relies on Bayesian hierarchical state-space models that include two or three main components. The first component describes the occupancy dynamics of taxa at locations defined by \code{var_id} from taxon- and location-specific probabilities of persistence and colonisation. Two other independent components describe the population growth of taxa on successively occupied locations, one based on abundance (if \code{var_cnt} is supplied) and other based on biomass (if \code{var_wei} is supplied). These two components are based on an autoregressive growth model. At the location level, the finite population growth rates are only estimated when the same sampling protocol is carried out from one time to the next from the changes in population size. At global level (i.e. across all locations) and regional levels defined by \code{var_reg}, the finite population growth rates are derived from the size-weighted average of the finite growth rates estimated at the location level. The products and the geometric means of the finite population growth rates are respectively used as estimates of growth rates and average growth rates over the whole time modelling period and over intermediate time periods defined by \code{period}. If \code{var_guild} is supplied, the guild growth rates are assessed from the geometric mean of taxa growth rates.  

The models are run using the \code{runMCMC} function of \code{nimble} package.
}
\value{
\code{mod_popgrowAlt} returns a list with the summaries of main statistics associated with the estimation of the requested parameters \code{(mcmc_summary}), the MCMC chains (\code{mcmc_chain}) and the list of parameters subscripts (\code{subscript}).
}
\seealso{
\code{\link{mod_popgrow}} and \code{\link{get_modparameters}}
}
\description{
Fits Bayesian hierarchical single or multi-species population growth models to abundance and/or biomass time series. The modelling framework not accounts for detection biases of taxa and individuals but allows changes in the sampling protocol to be considered.
}
\examples{
\dontrun{
df <- riverfish[riverfish$pass == 1,]

# Pike population growth from abundance
mcmc.out <- mod_popgrowAlt(df[df$taxa == "pike",],
                           var_id = "pop_id",
                           var_tmp = "year",
                           var_cnt = "headcount",
                           var_pro = "fishing")
                           
# Multi-taxa population growth from biomass                           
mcmc.out <- mod_popgrowAlt(df,
                           var_id = "pop_id",
                           var_tmp = "year",
                           var_tax = "taxa",
                           var_wei = "biomass",
                           var_pro = "fishing")
                           
# Multi-taxa population growth from abundance and biomass                           
mcmc.out <- mod_popgrowAlt(df,
                           var_id = "pop_id",
                           var_tmp = "year",
                           var_tax = "taxa",
                           var_cnt = "headcount",
                           var_wei = "biomass",
                           var_pro = "fishing")
}
}
