% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popoccup.R
\name{mod_popoccup}
\alias{mod_popoccup}
\title{Runs occupancy models in a Bayesian hierarchical framework}
\usage{
mod_popoccup(
  df,
  var_id,
  var_tmp,
  var_cnt,
  var_tax = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 10000,
  n_thin = ceiling(n_iter/100),
  n_burnin = floor(n_iter/4)
)
}
\arguments{
\item{df}{data frame}

\item{var_id}{character}

\item{var_tmp}{character}

\item{var_cnt}{character}

\item{var_tax}{character}

\item{var_reg}{character or character vector}

\item{var_guild}{character or character vector}

\item{period}{list}

\item{timestep}{numeric}

\item{save_parameters}{character or character vector}

\item{n_chain}{numeric}

\item{n_iter}{numeric}

\item{n_thin}{numeric}

\item{n_burnin}{numeric}
}
\value{
output list
}
\description{
Runs occupancy models in a Bayesian hierarchical framework
}
\examples{
\dontrun{
data(riverfish)
df <- riverfish[riverfish$pass == 1,]
mcmc.out <- mod_popoccup(df[df$taxa == "eel",],
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
n_chain=3, n_iter=10, n_thin=1, n_burnin=0)
mcmc.out <- mod_popoccup(df,
var_id = "pop_id",
var_tmp = "year",
var_tax = "taxa",
var_cnt = "headcount",
var_reg = "hydro_basin",
var_guild = "family",
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- mod_popoccup(df,
var_id = "pop_id",
var_tmp = "year",
var_tax = "taxa",
var_cnt = "headcount",
var_reg = "hydro_basin",
var_guild = "family",
period = list(c(2010,2016),c(2016,2022)),
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
}
}
