% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modenv_popgrow.R
\name{modenv_popgrow}
\alias{modenv_popgrow}
\title{Runs multi-taxa environmental growth models in a Bayesian hierarchical framework}
\usage{
modenv_popgrow(
  df,
  var_id,
  var_tmp,
  var_env,
  var_tax = NULL,
  var_cnt = NULL,
  var_wei = NULL,
  var_surf = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 10000,
  n_thin = ceiling(n_iter/100),
  n_burnin = floor(n_iter/4)
)
}
\arguments{
\item{df}{data frame}

\item{var_id}{character}

\item{var_tmp}{character}

\item{var_env}{character or character vector}

\item{var_tax}{character}

\item{var_cnt}{character}

\item{var_wei}{character}

\item{var_surf}{character}

\item{var_reg}{character or character vector}

\item{var_guild}{character or character vector}

\item{period}{list}

\item{timestep}{numeric}

\item{save_parameters}{character or character vector}

\item{n_chain}{numeric}

\item{n_iter}{numeric}

\item{n_thin}{numeric}

\item{n_burnin}{numeric}
}
\value{
output list
}
\description{
Runs multi-taxa environmental growth models in a Bayesian hierarchical framework
}
\examples{
\dontrun{
data(riverfish)
df <- riverfish[riverfish$pass == 1,]
mcmc.out <- modenv_popgrow(df,
var_id = "pop_id",
var_tmp = "year",
var_tax = "taxa",
var_env = "temperature",
var_cnt = "headcount",
var_surf = "surface",
var_guild = c("family","gudgeon"),
n_chain=3,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- modenv_popgrow(df[df$taxa == "pike",],
var_id = "pop_id",
var_tmp = "year",
var_env = "temperature",
var_cnt = "headcount",
var_wei = "biomass",
var_surf = "surface",
var_reg = "hydro_basin",
period = list(c(2010,2016),c(2016,2022)),
save_parameters = c("N_mulambda","B_mulambda"),
n_chain=3,n_iter=10,n_thin=1,n_burnin=0)
}
}
