% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modenv_popoccup.R
\name{modenv_popoccup}
\alias{modenv_popoccup}
\title{Runs multi-taxa environmental occupancy models in a Bayesian hierarchical framework}
\usage{
modenv_popoccup(
  df,
  var_id,
  var_tmp,
  var_cnt,
  var_tax = NULL,
  var_envO = NULL,
  var_envP = NULL,
  var_envC = NULL,
  env_tmp = TRUE,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 10000,
  n_thin = ceiling(n_iter/100),
  n_burnin = floor(n_iter/4)
)
}
\arguments{
\item{df}{data frame}

\item{var_id}{character}

\item{var_tmp}{character}

\item{var_cnt}{character}

\item{var_tax}{character}

\item{var_envO}{character or character vector}

\item{var_envP}{character or character vector}

\item{var_envC}{character or character vector}

\item{env_tmp}{logical}

\item{var_reg}{character or character vector}

\item{var_guild}{character or character vector}

\item{period}{list}

\item{timestep}{numeric}

\item{save_parameters}{character or character vector}

\item{n_chain}{numeric}

\item{n_iter}{numeric}

\item{n_thin}{numeric}

\item{n_burnin}{numeric}
}
\value{
output list
}
\description{
Runs multi-taxa environmental occupancy models in a Bayesian hierarchical framework
}
\examples{
\dontrun{
data(riverfish)
df <- riverfish[riverfish$pass == 1,]
mcmc.out <- modenv_popoccup(df[df$taxa == "eel",],
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
var_envO = c("temperature","depth"),
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- modenv_popoccup(df[df$taxa == "eel",],
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
var_envO = c("temperature","depth"),
var_reg = "hydro_basin",
period = list(c(2010,2016),c(2016,2022)),
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- modenv_popoccup(df,
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
var_tax = "taxa",
var_envP = c("temperature","depth"),
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- modenv_popoccup(df,
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
var_tax = "taxa",
var_envP = c("temperature","depth"),
var_envC = "temperature",
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
mcmc.out <- modenv_popoccup(df,
var_id = "pop_id",
var_tmp = "year",
var_cnt = "headcount",
var_tax = "taxa",
var_envP = c("temperature","depth"),
var_envC = "temperature",
var_guild = "family",
n_chain=2,n_iter=10,n_thin=1,n_burnin=0)
}
}
