% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverfish.R
\docType{data}
\name{modfunparameters}
\alias{modfunparameters}
\title{Parameters of modelling functions}
\description{
This data frame gives the description of the parameters used in modelling functions.
}
\usage{data(modfunparameters)}
\format{
  \code{modfunparameters} is a data frame with 4 variables:
  \describe{
    \item{\code{Para}}{parameter name}
    \item{\code{Function}}{function name}
    \item{\code{Parameter}}{parameter notation}
    \item{\code{Description}}{parameter description}
   }
}
\keyword{datasets}