% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popbay_data.R
\encoding{UTF-8}
\name{popbay_data}
\alias{popbay_data}
\title{
Checking and standardising datasets for \code{popbay_model}
}
\description{
\code{popbay_data} checks and standardises a dataset in the format requested by \code{\link{popbay_model}}. 
}
\usage{
popbay_data(data, ColNames, time.interval, 
  start, end, na.replace)
}
\arguments{
  \item{data}{
a data frame containing the items (columns) used by \code{popbay_model} and possible additional items. 
}
  \item{ColNames}{
an optional list that defines the new names to be assigned to items to satisfy the names requested by \code{popbay_model}.
}
  \item{time.interval}{
an optional numeric argument that defines time interval between two consecutive dates (default is 1).
}
  \item{start}{
an optional numeric argument that defines the starting date of the modelling period (defaut is the minimum data value). 
}
  \item{end}{
an optional numeric argument that defines the ending date of the modelling period (defaut is the maximum data value). 
}
  \item{na.replace}{
an optional logical argument, if TRUE the missing values in surface and pro_id items are imputed (defaut is FALSE). 
}
}
\details{
\code{popbay_data} standardises the column names of an input data frame and checks if the data frame items match with the \code{popbay_model} requirements: variable format, missing dates, duplicate entries with different values (duplicated rows are removed), and missing values. If no error is detected and \code{na.replace} is TRUE, the function imputes the missing values in sampling surfaces (\code{surface}) and protocols (\code{pro_id}). At the site-base scale, the sampling surface and protocol of the nearest previous non-missing entries are used for imputation. \code{popbay_data} also displays warning messages if some entries do not meet the model conditions. 
}
\value{
\code{popbay_data} returns a data frame.
}
\seealso{\code{\link{popbay_model}}}
\examples{
\donttest{data <- data.frame("year"=rep(2010:2020,4),
  "region"=rep(c("Occitanie","PACA"),each=22),
  "site"=rep(c("S1","S2","S3","S4"),each=11),
  "abundance"=sample(0:100,44))

new.data <- popbay_data(data,
  ColNames=list(sta_id="site",date="year",bas_id="region"))
}
}
