% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popbay_model.R
\encoding{UTF-8}
\name{popbay_model}
\alias{popbay_model}
\title{
Fitting hierarchical Bayesian population dynamics models
}
\description{
\code{popbay_model} is used to write and fit a population dynamics model to occupancy, abundance and/or biomass time series in a hierarchical Bayesian framework.
}
\usage{
popbay_model(data, type, error,
  time.interval, start, end, period,
  parameters.to.save, n.chains, n.iter, n.burnin, n.thin,
  fit.model, file.model, file.mcmc)
}
\arguments{
  \item{data}{
a data.frame containing the variables used in the model and possible additional items (variables order does not matter):
\itemize{
\item sta_id: sampling site identifier (required numeric or character) 
\item date: sampling date (required numeric)
\item sample: sample position in multiple-pass removal sampling (numeric required if error is TRUE, must not contain missing values)
\item abundance: number of sampling individuals (numeric required for occupancy and abundance growth models)
\item biomass: weight of sampling individuals (numeric required for biomass growth models)
\item surface: sampling surface (numeric required for population growth models, must not contain missing values) 
\item pro_id: sampling protocol (numeric or character required for population growth models and multiple-pass sampling-based occupancy model, must not contain missing values) 
\item bas_id: region identifier (optional numeric or character)
}}
  \item{type}{
an optional character that defines the modelling type (default value is A): 
\itemize{
\item A: modelling population growth in abundance
\item B: modelling population growth in biomass
\item AB: modelling population growth in abundance and biomass
\item O: modelling distribution changes
\item OA: modelling distribution changes and population growth in abundance
\item OB: modelling distribution changes and population growth in biomass
\item OAB: modelling distribution changes and population growth in abundance and biomass
}}
  \item{error}{
an optional logical, if TRUE the modelling framework accounts for biases in the detection of individuals by removal sampling, if FALSE these biases are not taken into account (default value is FALSE).
}
  \item{time.interval}{
an optional numeric that defines the time interval between two consecutive dates (default is 1).
}
  \item{start}{
an optional numeric that defines the starting date for estimating average rates (defaut is the minimum data value)
}
  \item{end}{
an optional numeric that defines the ending date for estimating average rates (defaut is the maximum data value)
}
  \item{period}{
a numeric that defines the time interval for estimating intermediate average rates (defaut is NA, intermediate average rates are not estimated).
}
  \item{parameters.to.save}{
an optional character vector of the names of the parameters to save (see Details). By default, the function saves the average rates at site and national scales, at regional scale (if \code{bas_id} is specified in the data), the intermediate average rates (if \code{period} is specified) and the parameters related to detection errors (if \code{error} is TRUE).
}
  \item{n.chains}{
an optional numeric, number of Markov chains (default is 3).
}
  \item{n.iter}{
an optional numeric, total number of iterations per Markov chain (default is 102,000).
}
  \item{n.burnin}{
an optional numeric, number of iterations to discard at the beginning of Markov chains (default is 2000).
}
  \item{n.thin}{
an optional numeric, backup interval of iterations (defaut is 1000).
}
  \item{fit.model}{
an optional logical, if FALSE the model variables list and the model file are created, but the model is not run (defaut is TRUE).
}
  \item{file.model}{
an optional character that defines the location and/or the name of output files for saving the model file and the model variables list (the model is saved in .txt format and the model data list is saved in .rda format). If not specified, the model file and the model variables list are not saved.
}
  \item{file.mcmc}{
an optional character that defines the location and/or the name of an output file for saving the Markov chains (saved in .rda format). If not specified, the Markov chains are not saved.
}}
\details{
\code{popbay_model} relies on hierarchical Bayesian state-space models that can include one or more of three main components. The first component, systematically included, describes the sites occupancy dynamics over time from presence-absence data. Two independant components describe the population dynamics on successively occupied sites, one based on abundance and other based on biomass and allow for estimating growth rates at site-based scales. Either, both, or neither of these two components may be included in the model. Additional components allow for estimating distribution change rates and growth rates at regional and national scales.

The modelling framework requires complete time series. If \code{error} is TRUE, the modelling framework account for biases in species detection and individuals detection using a multiple-pass sampling method with removal samples for individuals detection. If \code{error} is FALSE, the modelling framework not accounts for detection biases and the growth modelling only supports sites sampled using the same protocol or two different protocols with a single protocol change in time.

The models are run using the \code{\link[R2jags]{jags}} function of R2jags package.

Occupancy sub-model parameters:
\itemize{
\item Ro.nat: average distribution change rate at national level
\item ro.nat: inter-time distribution change rates at national level
\item Ro.nat.period: intermediate average distribution change rate at national level
\item Ro.bas: average distribution change rate at regional level
\item ro.bas: inter-time distribution change rates at regional level
\item Ro.bas.period: intermediate average distribution change rate at regional level
\item p.per: site-specific persistence probability
\item p.col: site-specific colonisation probability
\item p.cap.sp: species detection probability specific to site and sampling protocol (if \code{error} is TRUE)
\item z: true occupancy states (if \code{error} is TRUE) or observed occupancy states (if \code{error} is FALSE) of sites
}

Abundance growth sub-model parameters:
\itemize{
\item R.nat: national average growth rate
\item r.nat: inter-time growth rates at national level
\item R.nat.period: intermediate national average growth rate
\item R.bas: regional average growth rate
\item r.bas: inter-time growth rates at regional level
\item R.bas.period: intermediate regional average growth rate
\item R.sta: site average growth rate
\item r.sta: inter-time growth rates at site level
\item R.sta.period: intermediate average growth rate at site level
\item nb.r.sta: number of inter-time growth rates estimated at site level
\item y: observed site abundances with imputation of missing values (if \code{error} is FALSE)
\item p.cap: individuals detection probability specific to site and sampling protocol (if \code{error} is TRUE)
\item N: true site abundances (if \code{error} is TRUE)
\item Nb: true regional abundances (if \code{error} is TRUE)
\item Nn: true national abundance (if \code{error} is TRUE)
}

Biomass growth sub-model parameters:
\itemize{
\item Rw.nat: national average growth rate
\item rw.nat: inter-time growth rates at national level
\item Rw.nat.period: intermediate national average growth rate
\item Rw.bas: regional average growth rate
\item rw.bas: inter-time growth rates at regional level
\item Rw.bas.period: intermediate regional average growth rate
\item Rw.sta: site average growth rate
\item rw.sta: inter-time growth rates at site level
\item Rw.sta.period: intermediate average growth rate at site level
\item nb.rw.sta: number of inter-time growth rates estimated at site level
\item w: observed site biomasses with imputation of missing values (if \code{error} is FALSE)
\item var.cap: variance of individuals detection error specific to site and sampling protocol (if \code{error} is TRUE)
\item B: true site biomasses (if \code{error} is TRUE)
\item Bb: true regional biomasses (if \code{error} is TRUE)
\item Bn: true national biomass (if \code{error} is TRUE)
}
}
\value{
\code{popbay_model} returns a data frame summarising the main statistics associated with the estimation of the requested parameters (mean, standard deviation, 2.5, 25, 50, 75, and 97.5 percent quantiles), the Brooks-Gelman-Rubin statistic (Rhat) and the effective sample size (n.eff).

If \code{fit.model} is FALSE, \code{popbay_model} returns a character vector with the names of the requested parameters.
}
\seealso{
\code{\link{popbay_data}}, help of \code{\link[R2jags]{jags}} for more information on the model run
}
\examples{
\donttest{data <- data.frame("year"=rep(2010:2020,4),
  "region"=rep(c("Occitanie","PACA"),each=22),
  "site"=rep(c("S1","S2","S3","S4"),each=11),
  "abundance"=sample(0:100,44))
# check data  
new.data <- popbay_data(data,
  ColNames=list(sta_id="site",date="year",bas_id="region"))
# create 'surface' and 'pro_id' variables  
new.data$surface <- 1
new.data$pro_id <- 1
# run the model
stat <- popbay_model(new.data,n.iter=1000,n.burnin=10,n.thin=10)
}
\dontrun{
## Abundance and biomass dynamics models 
data(riverfish)
stat <- popbay_model(riverfish,
  type="AB",
  n.iter=1000,n.burnin=10,n.thin=10)
# estimation of intermediate rates every 6 years:
stat <- popbay_model(riverfish,
  type="AB",
  period=6,
  n.iter=1000,n.burnin=10,n.thin=10)
# on-site survey every 2 years:
stat <- popbay_model(riverfish,
  type="AB",
  time.interval=2,
  n.iter=1000,n.burnin=10,n.thin=10)
# different starting date:
stat <- popbay_model(riverfish,
  type="AB",
  start=2010,
  n.iter=1000,n.burnin=10,n.thin=10)
# accounting for detection errors:
stat <- popbay_model(riverfish,
  type="AB",
  error=TRUE,
  n.iter=1000,n.burnin=10,n.thin=10)
  
## Write and save the variables and the model files 
# without running the model
para <- popbay_model(riverfish,
  type="OAB",
  error=TRUE,
  fit.model=FALSE,
  file.model="complete")

load("complete_datamodel.rda")
stat <- jags(data = data.jags, 
  parameters.to.save=para, 
  model.file = "complete_model.txt")
}
}
\keyword{Bayesian}
\keyword{modelling}