% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynmodel-package.R
\docType{package}
\name{popdynmodel-package}
\alias{popdynmodel}
\alias{popdynmodel-package}
\title{popdynmodel: an R Package to implement population dynamics models in hierarchical Bayesian framework}
\description{
This package provides functions to write and implement population dynamics models in a hierarchical Bayesian framework. Models are intended for occupancy, abundance and/or biomass time series and allow for estimations of occupancy changes rates and population growth rates at various spatial and temporal scales.
}
\author{
\strong{Maintainer}: Dortel E. \email{emmanuelle.dortel@gmail.com}
}
\keyword{internal}