% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverfish.R
\docType{data}
\name{riverfish}
\alias{riverfish}
\title{Fish monitoring time series}
\description{
This dataset gives the abundance and biomass of seven freshwater fish species on 85 sites
sampled, from 2010 to 2022, by multiple-pass removal electrofishing. The sites are spread
over six river-basins of metropolitan France.
}
\usage{data(riverfish)}
\format{
  \code{riverfish} is a data frame with 14 variables:
  \describe{
    \item{\code{pop_id}}{sampling site identifier}
    \item{\code{year}}{sampling year}
    \item{\code{taxa}}{species common names}
    \item{\code{family}}{species family names}
    \item{\code{gudgeon}}{gudgeon guild: value is 1 if the species is a gudgeon and NA otherwise}
    \item{\code{pass}}{pass position in multi-pass removal electrofishing}
    \item{\code{headcount}}{fishing catch size by species}
    \item{\code{biomass}}{fishing catch weight by species}
    \item{\code{abundance}}{fishing catch size (numeric)}
    \item{\code{biomass}}{fishing catch weight (numeric)}
    \item{\code{surface}}{fishing surface}
    \item{\code{fishing}}{electrofishing protocol}
    \item{\code{hydro_basin}}{river-basin identifiers}
    \item{\code{temperature}}{water temperature per site and year}
    \item{\code{depth}}{fishing depth per site and year}
    \item{\code{flow}}{watercourse flow per site and year}
  }
}
\source{
\code{riverfish} is based on data extracted from the national ASPE database managed by the
French Biodiversity Agency (OFB).
}
\keyword{datasets}
\keyword{ASPE}