% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wri_MCMCsummary.R
\name{wri_MCMCsummary}
\alias{wri_MCMCsummary}
\title{Summary statistics from MCMC chains}
\usage{
wri_MCMCsummary(mcmc_chain, pop_internal)
}
\arguments{
\item{mcmc_chain}{MCMC output in mcmc or mcmc.list format}

\item{pop_internal}{list with internal information linked to input model data (see \code{\link{wri_popmodel}})}
}
\value{
\code{wri_MCMCsummary} returns a list with the summaries of main statistics associated with the estimation of model parameters (\code{mcmc_summary}), the MCMC chains after removing the missing values (\code{mcmc_chain}) and the list of parameters subscripts (\code{subscript}).
}
\description{
Extracts summary statistics from one or more MCMC chains.
}
\examples{
\dontrun{
## Write multi-taxa occupancy model
df <- riverfish[riverfish$pass == 1,]
modpopoccup <- wri_popmodel(df, 
                            model = "mod_popoccup",
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_tax = "taxa",
                            var_cnt = "headcount")
                            
## Run the model using the nimble package
nimModel <- nimbleModel(code = modpopoccup$pop_code,
                        constants = modpopoccup$pop_constants,
                        data = modpopoccup$pop_data,
                        inits = modpopoccup$pop_inits,
                        name = "nimModel", calculate = FALSE)
MCMCConf <- configureMCMC(nimModel, monitors = modpopoccup$pop_parameters)
MCMCbuild <- buildMCMC(MCMCConf)
MCMCComp <- compileNimble(nimModel)
nimModelComp <- compileNimble(MCMCbuild, project = nimModel, resetFunctions = TRUE)
mcmc_chain <- runMCMC(nimModelComp, 
                      nchains = 3, niter = 100, thin = 1, nburnin = 10,
                      samplesAsCodaMCMC = TRUE)

## Extract summary statistic
MCMC_summary <- wri_MCMCsummary(mcmc_chain, modpopoccup$pop_internal)  
}
}