% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wri_popmodel.R
\name{wri_popmodel}
\alias{wri_popmodel}
\title{Population model code and corresponding input data files}
\usage{
wri_popmodel(
  df,
  model,
  var_id,
  var_tmp,
  var_tax = NULL,
  var_cnt = NULL,
  var_wei = NULL,
  var_surf = NULL,
  var_reg = NULL,
  var_guild = NULL,
  var_env = NULL,
  var_envO = NULL,
  var_envP = NULL,
  var_envC = NULL,
  var_pro = NULL,
  var_pas = NULL,
  var_det = NULL,
  period = NULL,
  timestep = 1
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{model}{[character] expected population model (\code{mod_pop...} or \code{modenv_pop...})}

\item{var_id}{[character] variable name with sampling location identifiers (must not contain missing values)}

\item{var_tmp}{[character] variable name with continuous sampling times (must be numeric without missing values)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{var_cnt}{[optional, character] variable name with taxa abundances (must be numeric)}

\item{var_wei}{[optional, character] variable name with taxa biomasses (must be numeric)}

\item{var_surf}{[optional, character] variable name with quantitative sampling effort (e.g. sampling surfaces). If supplied, the model takes into account variabilities in sampling effort from one time to the next (must be numeric without missing values)}

\item{var_reg}{[optional, character] variable name(s) with spatial level identifiers. If supplied, estimates of occupancy changes rates at intermediate spatial levels are added to the model. Otherwise, occupancy change rates are only estimated across all locations}

\item{var_guild}{[optional, character] variable name(s) with taxa guilds identifiers. If supplied, estimates of guild occupancy changes rates are added to the model}

\item{var_env}{[optional, character] environmental covariate(s) name(s) related to growth rates whose values must be supplied by location and/or time (must be numeric without missing values)}

\item{var_envO}{[optional, character] environmental covariate(s) name(s) related to occupancy probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envP}{[optional, character] environmental covariate(s) name(s) related to persistence probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envC}{[optional, character] environmental covariate(s) name(s) related to colonisation probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_pro}{[optional, character] variable name with sampling protocol (must not contain missing values)}

\item{var_pas}{[optional, character] variable name with pass numbers in a multiple-pass removal sampling (must be numeric without missing values)}

\item{var_det}{[optional, character] environmental covariate(s) name(s) related to taxa detection probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{period}{[optional, numeric] list of vectors with start and end times for estimating intermediate average growth rates}

\item{timestep}{[optional, numeric] interval between two consecutive times to account for regular gaps in time series in the rate estimates}
}
\value{
\code{wri_popmodel} returns a list with the model code (\code{pop_code}), the input data in model format (\code{pop_data} and \code{pop_constants}), the initial values for model variables (\code{pop_inits}), the name of the model parameters (\code{pop_parameter}) and information used by the \code{wri_MCMCsummary} function (\code{pop_internal}). See example of \code{\link{wri_MCMCsummary}} function to execute the models.
}
\seealso{
\code{\link{wri_MCMCsummary}} and \code{\link{get_modparameters}}
}
\description{
Writes the expected population model code and the corresponding input date files in the format used by the \code{nimble} package.
}
\examples{
df <- riverfish[riverfish$pass == 1,] 

## Multi-taxa occupancy and population growth from abundance
modpopdyn <- wri_popmodel(df,
                          model = "mod_popdyn",
                          var_id = "pop_id",
                          var_tmp = "year",
                          var_tax = "taxa",
                          var_cnt = "headcount")
                          
## Alternative multi-taxa occupancy and population growth from abundance
modpopdynAlt <- wri_popmodel(df,
                             model = "mod_popdynAlt",
                             var_id = "pop_id",
                             var_tmp = "year",
                             var_tax = "taxa",
                             var_cnt = "headcount",
                             var_pro = "fishing")


## Eel population growth from abundance with linear temperature effect on growth rates
modenvpopgrow <- wri_popmodel(df[df$taxa \%in\% "eel",],
                              model = "modenv_popgrow",
                              var_id = "pop_id",
                              var_tmp = "year",
                              var_env = "temperature",
                              var_cnt = "headcount")
}