% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writemodel_unbiased.R
\encoding{UTF-8}
\name{writemodel_unbiased}
\alias{writemodel_unbiased}
\title{
Writing unbiased hierarchical Bayesian population dynamics models
}
\description{
\code{writemodel_unbiased} writes hierarchical Bayesian models for occupancy, abundance and/or biomass time series. The models not account for biases induced by imperfect individual detection. This function is used by the \code{\link{popbay_model}}.
}
\usage{
writemodel_unbiased(type, period,
  para.nat, para.bas, para.sta)
}
\arguments{
  \item{type}{
a character that defines the modelling type:
\itemize{
\item A: modelling population growth in abundance
\item B: modelling population growth in biomass
\item AB: modelling population growth in abundance and biomass
\item O: modelling distribution changes
\item OA: modelling distribution changes and population growth in abundance
\item OB: modelling distribution changes and population growth in biomass
\item OAB: modelling distribution changes and population growth in abundance and biomass
}}
  \item{period}{
a numeric or charactere, if NA the written model does not allows for estimating intermediate average rates. 
}
  \item{para.nat}{
a logical, if FALSE the written model does not allows for estimating national average rates.
}
  \item{para.bas}{
a logical, if FALSE the written model does not allows for estimating regional average rates.
}
  \item{para.sta}{
a logical, if FALSE the written model does not allows for estimating intermediate average rates at site-base scale.
}
}
\value{
\code{writemodel_unbiased} returns the model code as a character vector object. The model can then be directly run using \code{textConnection} or save and run from backup file.
}
\seealso{
\code{\link{writemodel_removal}}, \code{\link{popbay_model}}
}
\examples{
\donttest{model <- writemodel_unbiased(type="A",
  period=NA,
  para.nat=TRUE,
  para.bas=FALSE,
  para.sta=FALSE)
}
\dontrun{
# Directly run the model 
stat <- jags(data = varlist,
  parameters.to.save=c("R.nat"),
  model=textConnection(model))
# Run the model from backup file
write(model,"model.txt")
stat <- jags(data = varlist,
  parameters.to.save=c("R.nat"),
  model="model.txt") 
}
}
