% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modparameters.R
\name{get_modparameters}
\alias{get_modparameters}
\title{Gives description of the parameters of modeling functions}
\usage{
get_modparameters(fun = NULL, parameter = NULL)
}
\arguments{
\item{fun}{character}

\item{parameter}{character or character vector}
}
\description{
Gives description of the parameters of modeling functions
}
\examples{
get_modparameters(fun = "mod_popoccup")
get_modparameters(parameter = "z")
get_modparameters(parameter = c("z","N","B"))
}
