% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_convertquali.R
\name{mef_convertquali}
\alias{mef_convertquali}
\title{Converts quantitative variables into qualitative variables}
\usage{
mef_convertquali(df, var_quali)
}
\arguments{
\item{df}{data frame}

\item{var_quali}{[character] name(s) of qualitative variable(s) to convert}
}
\details{
\code{mef_convertquali} converts each quantitative variable by breaking it into multiple binary variables (one fewer than the number of modalities of qualitative variable).
}
\value{
\code{mef_convertquali} returns the data frame \code{df} with the new binary variables.
}
\description{
Converts one or more quantitative variables into qualitative variables.
}
\examples{
data(riverfish)

# Conversion of variable flow
df <- mef_convertquali(riverfish, var_quali="flow")
}
