% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_imputevalue.R
\name{mef_imputevalue}
\alias{mef_imputevalue}
\title{Imputes missing values in times series}
\usage{
mef_imputevalue(df, var_id, var_tmp, var_imp)
}
\arguments{
\item{df}{data frame}

\item{var_id}{argument}

\item{var_tmp}{argument}

\item{var_imp}{argument}
}
\value{
df data frame
}
\description{
Imputes missing values in times series
}
\examples{
data(riverfish)
n <- sample(2010:2022, 4)
riverfish[riverfish$year \%in\% n,"surface"] <- NA
df <- mef_imputevalue(riverfish, var_id=pop_id, var_tmp=year, var_imp=surface)
}
