% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_popoccup.R
\name{mod_popoccup}
\alias{mod_popoccup}
\title{Population occupancy modelling in a Bayesian hierarchical framework}
\usage{
mod_popoccup(
  df,
  var_id,
  var_tmp,
  var_cnt,
  var_tax = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 102000,
  n_thin = 1000,
  n_burnin = 2000
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{var_id}{[character] variable name with sampling location identifiers (must not contain missing values)}

\item{var_tmp}{[character] variable name with continuous sampling times (must be numeric without missing values)}

\item{var_cnt}{[character] variable name used to define presence/absence (must be numeric)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{var_reg}{[optional, character] variable name(s) with spatial level identifiers. If supplied, estimates of occupancy changes rates at intermediate spatial levels are added to the model. Otherwise, occupancy change rates are only estimated across all locations}

\item{var_guild}{[optional, character] variable name(s) with taxa guilds identifiers. If supplied, estimates of guild occupancy changes rates are added to the model}

\item{period}{[optional, numeric] list of vectors with start and end times for estimating intermediate average growth rates}

\item{timestep}{[optional, numeric] interval between two consecutive times to account for regular gaps in time series in the rate estimates}

\item{save_parameters}{[optional, character] name(s) of the parameter(s) to save (by default, all model parametes are saved)}

\item{n_chain}{[optional, numeric] number of MCMC chains}

\item{n_iter}{[optional, numeric] number of iterations per MCMC chain}

\item{n_thin}{[optional, numeric] backup interval for MCMC samples}

\item{n_burnin}{[optional, numeric] number of initial iterations to discard per Markov chain}
}
\details{
\code{mod_popoccup} relies on Bayesian hierarchical state-space models that describe the occupancy dynamics of one or more taxa from presence-absence data. The successive occupancy states of locations defined by \code{var_id} depend on taxon- and location-specific probabilities of persistence and colonisation. At the global level (i.e. across all locations) and at intermediate spatial levels defined by \code{var_reg}, the taxa occupancy changes rates are estimated from the ratio of the number of occuped locations from one time to the next. Their product and their geometric mean are respectively used as estimates of occupancy change rate and average occupancy change rate over the whole time modelling period and over intermediate time periods defined by \code{period}. If \code{var_guild} is supplied, the guild occupancy change rates are also assessed from the geometric mean of taxa occupancy change rates. 

The models are run using the \code{runMCMC} function of \code{nimble} package.
}
\value{
\code{mod_popoccup} returns a list with the summaries of main statistics associated with the estimation of the requested parameters (\code{mcmc_summary}), the MCMC chains (\code{mcmc_chain}) and the list of parameters subscripts (\code{subscript}).
}
\seealso{
\code{\link{mod_popoccupRS}} and \code{\link{get_modparameters}}
}
\description{
Fits Bayesian hierarchical occupancy models to presence-absence time series for estimating the occurrence of one or more taxa from colonisation and extinction probabilities. The modelling framework not accounts for taxa detection biases.
}
\examples{
\dontrun{
df <- riverfish[riverfish$pass == 1,]

## Occupancy dynamics of eel
mcmc.out <- mod_popoccup(df[df$taxa == "eel",],
                         var_id = "pop_id",
                         var_tmp = "year",
                         var_cnt = "headcount")

## Occupancy dynamics of taxa and families at basin level
mcmc.out <- mod_popoccup(df,
                         var_id = "pop_id",
                         var_tmp = "year",
                         var_tax = "taxa",
                         var_cnt = "headcount",
                         var_reg = "hydro_basin",
                         var_guild = "family")

## With intermediate rates every 6 years
mcmc.out <- mod_popoccup(df,
                         var_id = "pop_id",
                         var_tmp = "year",
                         var_tax = "taxa",
                         var_cnt = "headcount",
                         var_reg = "hydro_basin",
                         var_guild = "family",
                         period = list(c(2010,2016),c(2016,2022)))
}
}
