% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modenv_popdyn.R
\name{modenv_popdyn}
\alias{modenv_popdyn}
\title{Population dynamics modelling in a Bayesian hierarchical framework with environmental stochasticity}
\usage{
modenv_popdyn(
  df,
  var_id,
  var_tmp,
  var_tax = NULL,
  var_cnt = NULL,
  var_wei = NULL,
  var_env = NULL,
  var_envO = NULL,
  var_envP = NULL,
  var_envC = NULL,
  var_surf = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 102000,
  n_thin = 1000,
  n_burnin = 2000
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{var_id}{[character] variable name with sampling location identifiers (must not contain missing values)}

\item{var_tmp}{[character] variable name with continuous sampling times (must be numeric without missing values)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{var_cnt}{[optional, character] variable name with taxa abundances (must be numeric)}

\item{var_wei}{[optional, character] variable name with taxa biomasses (must be numeric)}

\item{var_env}{[optional, character] environmental covariate(s) name(s) related to growth rates whose values must be supplied by location and/or time (must be numeric without missing values)}

\item{var_envO}{[optional, character] environmental covariate(s) name(s) related to occupancy probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envP}{[optional, character] environmental covariate(s) name(s) related to persistence probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envC}{[optional, character] environmental covariate(s) name(s) related to colonisation probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_surf}{[optional, character] variable name with quantitative sampling effort (e.g. sampling surfaces). If supplied, the model takes into account variabilities in sampling effort from one time to the next (must be numeric without missing values)}

\item{var_reg}{[optional, character] variable name(s) with spatial level identifiers. If supplied, estimates of occupancy changes rates at intermediate spatial levels are added to the model. Otherwise, occupancy change rates are only estimated across all locations}

\item{var_guild}{[optional, character] variable name(s) with taxa guilds identifiers. If supplied, estimates of guild occupancy changes rates are added to the model}

\item{period}{[optional, numeric] list of vectors with start and end times for estimating intermediate average growth rates}

\item{timestep}{[optional, numeric] interval between two consecutive times to account for regular gaps in time series in the rate estimates}

\item{save_parameters}{[optional, character] name(s) of the parameter(s) to save (by default, all model parametes are saved)}

\item{n_chain}{[optional, numeric] number of MCMC chains}

\item{n_iter}{[optional, numeric] number of iterations per MCMC chain}

\item{n_thin}{[optional, numeric] backup interval for MCMC samples}

\item{n_burnin}{[optional, numeric] number of initial iterations to discard per Markov chain}
}
\details{
\code{modenv_popdyn} relies on Bayesian hierarchical state-space models that include two or three main components. The first component describes the occupancy dynamics of one or more taxa from presence-absence data (defined from \code{var_cnt} if is supplied or, failing that, from \code{var_wei}). When \code{var_envO}, \code{var_envP} or \code{var_envC} are supplied,  the successive occupancy states of locations defined by \code{var_id} depend on taxon-specific responses to environmental covariates (c.f. \code{\link{modenv_popoccup}}). Otherwise the successive occupancy states of locations depend on taxon- and location-specific probabilities of persistence and colonisation. At the global level (i.e. across all locations) and at intermediate spatial levels defined by \code{var_reg}, the taxa occupancy changes rates are estimated from the ratio of the number of occuped locations from one time to the next. Their product and their geometric mean are respectively used as estimates of occupancy change rate and average occupancy change rate over the whole time modelling period and over intermediate time periods defined by \code{period}. 

Two other independent components describe the population growth of taxa on successively occupied location, from on headcounts (if \code{var_cnt} is supplied) and/or biomass (if \code{var_wei} is supplied). When \code{var_env} is supplied the local population growth rates are modelled as a linear function of environmental covariates. Otherwise, the local population growth rates are estimated from the changes in population size from one time to the next, as the growth rates at global and regional levels. The products and the geometric means of the finite population growth rates are respectively used as estimates of growth rates and average growth rates over the whole time modelling period and over intermediate time periods defined by \code{period}. 

If \code{var_guild} is supplied, the occupancy change rates and growth rates of taxa guilds are assessed from the geometric means of taxa rates.

The models are run using the \code{runMCMC} function of \code{nimble} package.
}
\value{
\code{modenv_popdyn} returns a list with the summaries of main statistics associated with the estimation of the requested parameters (\code{mcmc_summary}), the MCMC chains (\code{mcmc_chain}) and the list of parameters subscripts (\code{subscript}).
}
\seealso{
\code{\link{modenv_popoccup}}, \code{\link{modenv_popgrow}}, \code{\link{modenv_popdynRS}} and \code{\link{get_modparameters}}
}
\description{
Fits Bayesian hierarchical single or multi-taxa occupancy and growth models with environmental stochasticity to abundance and/or biomass time series. The modelling framework not accounts for detection biases of taxa and individuals.
}
\examples{
\dontrun{
df <- riverfish[riverfish$pass == 1,]

# Pike population dynamics 
# with linear temperature effect on growth rates
mcmc.out <- modenv_popdyn(df[df$taxa == "pike",],
                          var_id = "pop_id",
                          var_tmp = "year",
                          var_cnt = "headcount",
                          var_env = "temperature",
                          var_surf = "surface")
                          
# with linear temperature effect on occupancy probability
mcmc.out <- modenv_popdyn(df[df$taxa == "pike",],
                          var_id = "pop_id",
                          var_tmp = "year",
                          var_cnt = "headcount",
                          var_envO = "temperature",
                          var_surf = "surface")
                          
# with linear temperature effect on growth rates and occupancy probability
mcmc.out <- modenv_popdyn(df[df$taxa == "pike",],
                          var_id = "pop_id",
                          var_tmp = "year",
                          var_cnt = "headcount",
                          var_env = "temperature",
                          var_envO = "temperature",
                          var_surf = "surface")
                          
# Population dynamics of gudgeon species
# with environmental effects on persistence and colonisation probabilities
mcmc.out <- modenv_popdyn(df[!is.na(df$gudgeon),],
                          var_id = "pop_id",
                          var_tmp = "year",
                          var_tax = "taxa",
                          var_cnt = "headcount",
                          var_wei = "biomass",
                          var_envP = c("temperature","depth"),
                          var_envC = "temperature",
                          var_surf = "surface",
                          var_guild = "gudgeon")
}
}
