% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modenv_popdyn.R
\name{modenv_popdyn}
\alias{modenv_popdyn}
\title{Title}
\usage{
modenv_popdyn(
  df,
  var_id,
  var_tmp,
  var_tax = NULL,
  var_cnt = NULL,
  var_wei = NULL,
  var_env = NULL,
  var_envO = NULL,
  var_envP = NULL,
  var_envC = NULL,
  var_surf = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 10000,
  n_thin = ceiling(n_iter/100),
  n_burnin = floor(n_iter/4)
)
}
\arguments{
\item{n_burnin}{}
}
\description{
Title
}
