% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modenv_popoccup.R
\name{modenv_popoccup}
\alias{modenv_popoccup}
\title{Population occupancy modelling in a Bayesian hierarchical framework with environmental stochasticity}
\usage{
modenv_popoccup(
  df,
  var_id,
  var_tmp,
  var_cnt,
  var_tax = NULL,
  var_envO = NULL,
  var_envP = NULL,
  var_envC = NULL,
  var_reg = NULL,
  var_guild = NULL,
  period = NULL,
  timestep = 1,
  save_parameters = NULL,
  n_chain = 3,
  n_iter = 10000,
  n_thin = ceiling(n_iter/100),
  n_burnin = floor(n_iter/4)
)
}
\arguments{
\item{df}{data frame with variables used in the model}

\item{var_id}{[character] variable name with sampling location identifiers (must not contain missing values)}

\item{var_tmp}{[character] variable name with continuous sampling times (must be numeric without missing values)}

\item{var_cnt}{[character] variable name used to define presence/absence (must be numeric)}

\item{var_tax}{[optional, character] variable name with taxa identifiers to be specified for the multi-taxa models (must not contain missing values)}

\item{var_envO}{[optional, character] environmental covariate(s) name(s) related to occupancy probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envP}{[optional, character] environmental covariate(s) name(s) related to persistence probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_envC}{[optional, character] environmental covariate(s) name(s) related to colonisation probabilities whose values must be supplied by location and/or by time (must be numeric without missing values)}

\item{var_reg}{[optional, character] variable name(s) with spatial level identifiers. If supplied, estimates of occupancy changes rates at intermediate spatial levels are added to the model. Otherwise, occupancy change rates are only estimated across all locations}

\item{var_guild}{[optional, character] variable name(s) with taxa guilds identifiers. If supplied, estimates of guild occupancy changes rates are added to the model}

\item{period}{[optional, numeric] list with start and end times for estimating intermediate average growth rates}

\item{timestep}{[optional, numeric] interval between two consecutive times to account for regular gaps in time series in the rate estimates}

\item{save_parameters}{[optional, character] name(s) of the parameter(s) to save (by default, all model parametes are saved)}

\item{n_chain}{[optional, numeric] number of MCMC chains}

\item{n_iter}{[optional, numeric] number of iterations per MCMC chain}

\item{n_thin}{[optional, numeric] backup interval for MCMC samples}

\item{n_burnin}{[optional, numeric] number of initial iterations to discard per Markov chain}
}
\details{
\code{modenv_popoccup} relies on Bayesian hierarchical state-space models that describe the occupancy dynamics of one or more taxa from presence-absence data. The successive occupancy states of locations defined by \code{var_id} depend on taxon-specific responses to environmental covariates. At the global level (i.e. across all locations) and at intermediate spatial levels defined by \code{var_reg}, the taxa occupancy changes rates are estimated from the ratio of the number of occuped location from one time to the next. Their geometric mean is used as average occupancy rate over the whole time modelling period and over intermediate time periods defined by \code{period}. If \code{var_guild} is supplied, the guild occupancy change rates are assessed from the geometric means of taxa occupancy rates.

If \code{var_envO} is supplied, the occupancy states of the locations are modelled from the linear effect of environmental covariates on occupancy probabilities.

If \code{var_envP} and \code{var_envC} are supplied, the occupancy states of the locations are modelled from the linear effect of environmental covariates on persistence and colonisation probabilities.

If only \code{var_envP} is supplied, the occupancy states of the locations are modelled from the linear effect of environmental covariates on persistence probabilities and from taxon- and location-specific colonisation probabilities.

If only \code{var_envC} is supplied, the occupancy states of the locations are modelled from the linear effect of environmental covariates on colonisation probabilities and from taxon- andlocation-specific persistence probabilities.

The models are run using the \code{runMCMC} function of \code{nimble} package.
}
\value{
\code{modenv_popoccup} returns a list with the summaries of main statistics associated with the estimation of the requested parameters (\code{mcmc_summary}), the MCMC chains (\code{mcmc_chain}) and the list of parameters subscripts (\code{subscript}).
}
\seealso{
\code{\link{modenv_popoccupRS}} and \code{\link{get_modparameters}}
}
\description{
Fits Bayesian hierarchical occupancy models to presence-absence time series for estimating the occurrence of one or more taxa from relationships with environmental covariates. The modelling framework not accounts for taxa detection biases.
}
\examples{
\dontrun{
data(riverfish)
df <- riverfish[riverfish$pass == 1,]

# Occupancy dynamics of eel with linear environmental effects on occupancy probability
mcmc.out <- modenv_popoccup(df[df$taxa == "eel",],
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_cnt = "headcount",
                            var_envO = c("temperature","depth"))

# With intermediate rates every 6 years at basin level
mcmc.out <- modenv_popoccup(df[df$taxa == "eel",],
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_cnt = "headcount",
                            var_envO = c("temperature","depth"),
                            var_reg = "hydro_basin",
                            period = list(c(2010,2016),c(2016,2022)))

# Multi-taxa occupancy dynamics 
# with linear environmental effects on persistence probability
mcmc.out <- modenv_popoccup(df,
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_cnt = "headcount",
                            var_tax = "taxa",
                            var_envP = c("temperature","depth"))

# Multi-taxa occupancy dynamics 
# with linear environmental effects on persistence and colonisation probabilities
mcmc.out <- modenv_popoccup(df,
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_cnt = "headcount",
                            var_tax = "taxa",
                            var_envP = c("temperature","depth"),
                            var_envC = "temperature")

# Add occupancy change rates of families
mcmc.out <- modenv_popoccup(df,
                            var_id = "pop_id",
                            var_tmp = "year",
                            var_cnt = "headcount",
                            var_tax = "taxa",
                            var_envP = c("temperature","depth"),
                            var_envC = "temperature",
                            var_guild = "family")
}
}
