% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynmodel-package.R
\docType{package}
\name{popdynmodel-package}
\alias{popdynmodel}
\alias{popdynmodel-package}
\title{popdynmodel: POPULATION DYNAMICS MODELLING}
\description{
Provides functions to write and implement population dynamics models in a hierarchical Bayesian framework.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
