% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynmodel-package.R
\docType{package}
\name{popdynmodel-package}
\alias{popdynmodel}
\alias{popdynmodel-package}
\title{popdynmodel: an R package to implement population dynamics models in hierarchical Bayesian framework}
\description{
This package provides tools and functions to implement population dynamics models in a hierarchical Bayesian framework. The models are intended for occupancy, abundance and/or biomass time series. They allow the estimation of occupancy change rates and population growth rates, for one or more taxa, at various spatial and temporal scales.

Function names start with specific prefixes:
\itemize{
\item \code{get_} : get information about modelling functions
\item \code{mef_} : format an input dataframe to meet with the models requirements
\item \code{mod_} : modelling function
\item \code{modenv_} : modelling function with environmental stochasticity
}
}
\author{
\strong{Maintainer}: Dortel E. \email{emmanuelle.dortel@gmail.com}
}
\keyword{internal}