% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdynmodel-package.R
\docType{package}
\name{popdynmodel-package}
\alias{popdynmodel}
\alias{popdynmodel-package}
\title{
Population dynamics modelling package 
}
\description{
This package was developed to analyse the river fish monitoring data from the ASPE database managed by the French Biodiversity Agency (OFB), but it can be used for a wide variety of taxa. It provides functions to write and implement population dynamics models in a hierarchical Bayesian framework. Models are intended for occupancy, abundance and/or biomass time series and allow for estimations of distribution change rates and population growth rates at various spatial and temporal scales.

This package has two main functions : \code{popbay_model} selects and prepares data, writes and fits population dynamics models to selected data and \code{popbay_data} checks and converts a dataset to the format requested by \code{popbay_model}. 
}
\author{
\strong{Maintainer}: Dortel E. \email{emmanuelle.dortel@gmail.com}
}
\keyword{Bayesian}
\keyword{model}
