% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverfish.R
\docType{data}
\name{riverfish}
\alias{riverfish}
\title{Fish monitoring time series}
\description{
This data set gives the abundance and biomass of Brown trout populations at 30 sites sampled by multiple-pass removal electrofishing from 2000 to 2018. The sites are spread over four river-basins of metropolitan France.
}
\usage{riverfish}
\format{
  \code{riverfish} is a data frame with 8 variables:
  \describe{
    \item{\code{sta_id}}{sampling site identifier (numeric)}
    \item{\code{date}}{sampling year (numeric)}
    \item{\code{sample}}{sample position in multiple-pass removal electrofishing (numeric)}
    \item{\code{abundance}}{fishing catch size (numeric)}
    \item{\code{biomass}}{fishing catch weight (numeric)}
    \item{\code{surface}}{fishing surface (numeric)}
    \item{\code{pro_id}}{electrofishing protocol identifier (numeric)}
    \item{\code{bas_id}}{river-basin identifier (numeric)}
  }
}
\source{
This dataset was extracted from the national ASPE database managed by the French Biodiversity Agency (OFB).
}
\keyword{datasets}
\keyword{ASPE}